refactor: 统一使用浮点精度支持CREDIT资源类型
- 删除旧的int版本calculateAvailableCount，统一为float64版本
- 重命名GetAvailableVIBECount为GetAvailableCount
- 更新TokenWithUsage.AvailableCount和CachedToken.Available为float64
- 修改所有日志调用使用logger.Float64替代logger.Int
- 支持精确的小数使用量显示（如0.26 credits）
- 完全基于CREDIT资源类型，移除VIBE兼容代码

Breaking Change: AvailableCount字段从int改为float64

+++ auth/token_selection_strategy.go
@@ -63,7 +63,7 @@ func (oss *OptimalSelectionStrategy) SelectToken(tokens map[string]*CachedToken)
 	if bestKey != "" {
 		logger.Debug("最优策略选择token",
 			logger.String("selected_key", bestKey),
-			logger.Int("available_count", bestToken.Available))
+			logger.Float64("available_count", bestToken.Available))
 	}
 
 	return bestKey
@@ -107,7 +107,7 @@ func (sss *SequentialSelectionStrategy) SelectToken(tokens map[string]*CachedTok
 			if cached.IsUsable() {
 				logger.Debug("顺序策略选择token（无顺序配置）",
 					logger.String("selected_key", key),
-					logger.Int("available_count", cached.Available))
+					logger.Float64("available_count", cached.Available))
 				return key
 			}
 		}
@@ -123,7 +123,7 @@ func (sss *SequentialSelectionStrategy) SelectToken(tokens map[string]*CachedTok
 			logger.Debug("顺序策略选择token",
 				logger.String("selected_key", currentKey),
 				logger.Int("index", sss.currentIndex),
-				logger.Int("available_count", cached.Available))
+				logger.Float64("available_count", cached.Available))
 			return currentKey
 		}
